Schema changes to make some entity fields consistent.
git-svn-id: https://develop.eng.uts.edu.au/projects/personal/sahara/svn/SchedServer/trunk@592 7ee224fa-57ca-4419-87f9-1f4314938d77

+++ DataAccess/src/au/edu/uts/eng/remotelabs/schedserver/dataaccess/dao/tests/UserDaoTester.java
@@ -160,7 +160,7 @@ public class UserDaoTester extends TestCase
         Session ses = DataAccessActivator.getNewSession();
         
         ses.beginTransaction();
-        UserClass uc = new UserClass("test1", 10, true, true, true, true);
+        UserClass uc = new UserClass("test1", (short) 10, true, true, true, true);
         ses.save(uc);
         User us = new User("del1", "ns1", "ACADEMIC");
         ses.save(us);
